requirements.txt
tabula-py
pandas
requests